refactor: Rename PhotoPicker to AvatarPicker
