+++ ui/src/lib/components/profiles/AvatarPicker.svelte
