+++ ui/src/routes/settings/profile/edit-photo/+page.svelte
@@ -8,7 +8,7 @@
 	import { Button, Page, Preloader } from 'konsta/svelte';
 	import { showToast } from '$lib/utils/toasts';
 	import { isIos } from '$lib/utils/environment';
-	import PhotoPicker from '$lib/components/profiles/PhotoPicker.svelte';
+	import AvatarPicker from '$lib/components/profiles/AvatarPicker.svelte';
 
 	const contactsStore: ContactsStore = getContext('contacts-store');
 	let avatar = $state<string | undefined>(undefined);
@@ -69,7 +69,7 @@
 		</div>
 	{:then myProfile}
 		<div class="column" style="flex: 1; overflow-y: auto;">
-			<PhotoPicker
+			<AvatarPicker
 				bind:avatar
 				bind:inModalState={inModalState}
 				onClose={() => goto('/settings/profile')}
